Completed basic simulator for basic testing. Added service booking function for medium level call

+++ src/Location.cpp
@@ -39,7 +39,7 @@ void Location::clearVehicleIds() {
 }
 
 int Location::getNumVehicles() {
-	return vehicle_ids_.size();
+	return (int) vehicle_ids_.size();
 }
 
 void Location::setCapacity(int capacity) {
